git 智能检测/project-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
scikit-learn
transformers
lime
requests
python-dotenv
httpx
celery
pika
opa-python-client
